pyproject.toml
[project]
name = "hackquest-soft"
version = "0.1.0"
description = ""
authors = [
    {name = "aionotaio",email = "143857085+aionotaio@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "better-proxy (>=1.3.1,<2.0.0)",
    "pyuseragents (>=1.0.5,<2.0.0)",
    "web3 (==6.20)",
    "pydantic (>=2.11.7,<3.0.0)",
    "inquirer (>=3.4.1,<4.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "wrapper-tls-requests (>=1.1.5,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
